+++ assets/select2.js
@@ -1,0 +1,6 @@
+/* global $ */
+
+window.addEventListener('load', function () {
+  const $selector = $('select.select2')
+  $selector.select2()
+})