+++ src/Controller/SidebarController.php
@@ -97,7 +97,7 @@ class SidebarController extends AbstractController
         if ($isCaseworker || $isAdministration) {
             $submenuItems[] = $this->generateSubmenuItem($this->translator->trans('Hearing', [], 'sidebar'), ['case_hearing_index', 'case_hearing_start', 'case_hearing_post_request_create', 'case_hearing_post_request_edit', 'case_hearing_post_response_create', 'case_hearing_post_show', 'case_hearing_post_response_edit'], $caseId, $activeRoute);
             // $submenuItems[] = $this->generateSubmenuItem($this->translator->trans('Communication', [], 'sidebar'), ['digital_post_index', 'digital_post_show'], $caseId, $activeRoute);
-            $submenuItems[] = $this->generateSubmenuItem($this->translator->trans('Case events', [], 'sidebar'), ['case_event_index', 'case_event_show', 'case_event_create', 'case_event_edit', 'digital_post_show'], $caseId, $activeRoute);
+            $submenuItems[] = $this->generateSubmenuItem($this->translator->trans('Case events', [], 'sidebar'), ['case_event_index', 'case_event_show', 'case_event_create', 'case_event_edit', 'case_event_copy', 'digital_post_show'], $caseId, $activeRoute);
 
             // Currently, they do not wish to create case presentation and decision proposal via TVIST1
             // $submenuItems[] = $this->generateSubmenuItem($this->translator->trans('Case presentation', [], 'sidebar'), ['case_presentation'], $caseId, $activeRoute);