+++ src/Service/CaseEventHelper.php
@@ -2,16 +2,19 @@
 
 namespace App\Service;
 
+use App\Entity\CaseDocumentRelation;
 use App\Entity\CaseEntity;
 use App\Entity\CaseEvent;
 use App\Entity\DigitalPost;
+use App\Entity\DigitalPostAttachment;
 use App\Entity\Party;
+use App\Repository\CaseEntityRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Security\Core\Security;
 
 class CaseEventHelper
 {
-    public function __construct(private EntityManagerInterface $entityManager, private Security $security)
+    public function __construct(private EntityManagerInterface $entityManager, private Security $security, private CaseEntityRepository $caseRepository)
     {
     }
 
@@ -20,7 +23,7 @@ class CaseEventHelper
         $caseEvent = new CaseEvent();
 
         $caseEvent
-            ->setCaseEntity($case)
+            ->addCaseEntity($case)
             ->setCategory(CaseEvent::CATEGORY_OUTGOING)
             ->setSubject(CaseEvent::SUBJECT_HEARING_CONTRADICTIONS_BRIEFING)
             ->setReceivedAt(new \DateTimeImmutable())
@@ -38,7 +41,7 @@ class CaseEventHelper
         $caseEvent = new CaseEvent();
 
         $caseEvent
-            ->setCaseEntity($case)
+            ->addCaseEntity($case)
             ->setCategory(CaseEvent::CATEGORY_NOTE)
             ->setSubject($subject)
             ->setReceivedAt($receivedAt)
@@ -57,7 +60,7 @@ class CaseEventHelper
         $caseEvent = new CaseEvent();
 
         $caseEvent
-            ->setCaseEntity($case)
+            ->addCaseEntity($case)
             ->setCategory(CaseEvent::CATEGORY_INCOMING)
             ->setSubject($subject)
             ->setReceivedAt($receivedAt)
@@ -84,5 +87,45 @@ class CaseEventHelper
     private function getLines(?string $additionalParties): array
     {
         return array_filter(array_map('trim', explode(PHP_EOL, $additionalParties)));
+    }
+
+    public function copyCaseEventToCases(array $cases, CaseEvent $caseEvent)
+    {
+        foreach ($cases as $case) {
+            $caseEvent->addCaseEntity($case);
+
+            // Handle documents
+            foreach ($caseEvent->getDocuments() as $document) {
+                $relation = new CaseDocumentRelation();
+                $relation->setCase($case);
+                $relation->setDocument($document);
+                $this->entityManager->persist($relation);
+            }
+
+            // Handle Digital Post
+            if ($digitalPost = $caseEvent->getDigitalPost()) {
+                $documents = array_map(static fn (DigitalPostAttachment $attachment) => $attachment->getDocument(), $digitalPost->getAttachments()->toArray());
+
+                // Add main document
+                $documents[] = $digitalPost->getDocument();
+
+                // Add attachments
+                foreach ($documents as $document) {
+                    $relation = new CaseDocumentRelation();
+                    $relation->setCase($case);
+                    $relation->setDocument($document);
+                    $this->entityManager->persist($relation);
+                }
+            }
+        }
+
+        $this->entityManager->flush();
+    }
+
+    public function findSuitableCasesForCopy(CaseEntity $case, CaseEvent $caseEvent, string $endStatus): array
+    {
+        $cases = $this->caseRepository->findNonFinishedCasesInSameBoard($case, $endStatus);
+
+        return array_diff($cases, $caseEvent->getCaseEntities()->toArray());
     }
 }