+++ webpack.config.js
@@ -38,6 +38,7 @@ Encore
   .addEntry('case_new_prepare', './assets/case/municipality_and_board_selector.js')
   .addEntry('mail_templates_custom_data', './assets/mail_templates/custom_data.js')
   .addEntry('document_case_event', './assets/document/case_event.js')
+  .addEntry('select2', './assets/select2.js')
 
   .addStyleEntry('address_protection', './assets/case/address-protection.scss')